requirements.txt
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.2.0
matplotlib==3.6.0
seaborn==0.12.0
streamlit==1.18.0
requests==2.28.0
python-dotenv==0.21.0
pytest==7.2.0
black==22.12.0
pylint==2.15.0
